requirements.txt
Django==4.0.4
django-split-settings==1.1.0
flake8==4.0.0
python-dotenv==0.20.0
django-debug-toolbar==3.4.0
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
requests==2.28.0
pandas==1.5.0
djangorestframework==3.14.0
celery==5.2.6
django-celery-results==2.4.0
django-celery-beat==2.4.0
redis==4.3.4
pytest==7.1.2
pytest-asyncio==0.12.0
django-rest-swagger